requirements.txt
browser-use
python-dotenv
langchain_google_genai
quart
pydantic